requirements.txt
beautifulsoup4==4.8.1
blis==0.4.1
catalogue==0.0.8
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cycler==0.10.0
cymem==2.0.3
de-core-news-md==2.2.5
decorator==4.4.1
dill==0.3.1.1
editdistance==0.5.3
geographiclib==1.50
geopy==1.20.0
idna==2.8
importlib-metadata==1.3.0
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.3
mip==1.6.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.0.2
multiprocess==0.70.9
murmurhash==1.0.2
networkx==2.4
numpy==1.17.4
pandas==0.25.3
plac==1.1.3
preshed==3.0.2
pycparser==2.19
Pympler==0.8
pyparsing==2.4.6
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
soupsieve==1.9.5
spacy==2.2.3
srsly==0.2.0
thinc==7.3.1
tqdm==4.40.2
urllib3==1.25.7
wasabi==0.4.2
zipp==0.6.0
